+++ src/transport/BUILD.gn
@@ -33,6 +33,14 @@ static_library("transport") {
     "SecureSessionMgr.h",
   ]
 
+  if (chip_config_network_layer_ble) {
+    sources += [
+      "BLE.cpp",
+      "BLE.h",
+    ]
+  }
+
+
   public_deps = [
     "${chip_root}/src/crypto",
     "${chip_root}/src/inet",
