+++ src/transport/raw/BUILD.gn
@@ -30,13 +30,6 @@ static_library("raw") {
     "UDP.h",
   ]
 
-  if (chip_config_network_layer_ble) {
-    sources += [
-      "BLE.cpp",
-      "BLE.h",
-    ]
-  }
-
   public_deps = [
     "${chip_root}/src/crypto",
     "${chip_root}/src/inet",
